LPS-61123 Move classes to internal package, no need to export
