+++ modules/portal/portal-cluster/src/main/java/com/liferay/portal/cluster/internal/BaseClusterReceiver.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.cluster;
+package com.liferay.portal.cluster.internal;
 
 import com.liferay.portal.kernel.cluster.Address;
 import com.liferay.portal.kernel.log.Log;
