+++ modules/portal/portal-cluster/src/main/java/com/liferay/portal/cluster/internal/ClusterForwardReceiver.java
@@ -14,7 +14,6 @@
 
 package com.liferay.portal.cluster.internal;
 
-import com.liferay.portal.cluster.BaseClusterReceiver;
 import com.liferay.portal.kernel.cluster.Address;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
