+++ modules/portal/portal-cluster/src/main/java/com/liferay/portal/cluster/internal/ClusterRequestReceiver.java
@@ -14,8 +14,6 @@
 
 package com.liferay.portal.cluster.internal;
 
-import com.liferay.portal.cluster.BaseClusterReceiver;
-import com.liferay.portal.cluster.ClusterChannel;
 import com.liferay.portal.kernel.cache.thread.local.Lifecycle;
 import com.liferay.portal.kernel.cache.thread.local.ThreadLocalCacheManager;
 import com.liferay.portal.kernel.cluster.Address;
