+++ modules/portal/portal-cluster/src/main/java/com/liferay/portal/cluster/internal/jgroups/JGroupsClusterChannel.java
@@ -14,8 +14,8 @@
 
 package com.liferay.portal.cluster.internal.jgroups;
 
-import com.liferay.portal.cluster.ClusterChannel;
-import com.liferay.portal.cluster.ClusterReceiver;
+import com.liferay.portal.cluster.internal.ClusterChannel;
+import com.liferay.portal.cluster.internal.ClusterReceiver;
 import com.liferay.portal.kernel.cluster.Address;
 import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.portal.kernel.io.Serializer;
