+++ modules/portal/portal-cluster/src/test/java/com/liferay/portal/cluster/internal/TestClusterChannel.java
@@ -14,8 +14,6 @@
 
 package com.liferay.portal.cluster.internal;
 
-import com.liferay.portal.cluster.ClusterChannel;
-import com.liferay.portal.cluster.ClusterReceiver;
 import com.liferay.portal.kernel.cluster.Address;
 import com.liferay.portal.kernel.util.ObjectValuePair;
